backend/requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10          # PostgreSQL driver
sentence-transformers>=2.6.0
redis>=5.0.0
# Auth
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
python-jose[cryptography]==3.3.0  # alternative JWT with more algorithms

# Email
aiosmtplib==3.0.2                 # async SMTP for email verification
jinja2==3.1.5                     # email templates

# RAG Pipeline
langchain==0.3.24
langchain-openai==0.3.7
langchain-community==0.3.21
langchain-huggingface==0.1.2
sentence-transformers==4.1.0
faiss-cpu==1.11.0

# Document Processing
PyMuPDF==1.25.5
python-docx==1.1.2

# HTTP Client (for calling STT/TTS services)
httpx==0.28.1

# WebSocket
websockets==15.0.1

# Rate limiting
slowapi==0.1.9
